Move everything into one file for now
